fw_pos_control_l1 extract FixedwingPositionControl header
